+++ src_util/ft_str4.c
@@ -1,9 +1,9 @@
 #include "minishell.h"
 
-int	ft_rmv_char(char **buffadr, int i)
+int	ft_rmv_char(char **buffadr, size_t i)
 {
-	int		x;
-	int		j;
+	size_t	x;
+	size_t	j;
 	size_t	size;
 	char	*str;
 	char	*tmp;
@@ -14,14 +14,13 @@ int	ft_rmv_char(char **buffadr, int i)
 		return (-1);
 	x = 0;
 	j = 0;
-	while (x < size-1)
+	while (x < size - 1)
 	{
 		if (x == i)
 		{
 			j++;
 		}
 		tmp[x] = str[j];
-		//printf("{--%s--}[%d]{%d}{%d}{%d} char : {%c}\n", str, i, size, x, j, str[j]);
 		x++;
 		j++;
 	}
